Module Hpack.Decoder

type t

val create : int -> t

create capacity initializes a decoder with a dynamic table with maximum size capacity. This size is an approximation of the memory usage in bytes.

See RFC75414.1 for more details.

val set_capacity : t -> int -> (unit, error) result

set_capacity decoder capacity sets decoder's dynamic table size to maximum size capacity. This size is an approximation of the memory overhead in bytes.

See RFC75406.5.2 and RFC75414.1 for more details.

val decode_headers : t -> (header list, error) result Angstrom.t

decode_headers decoder creates an Angstrom parser that will decode a header block and return a list of the decoded headers