Module Hpack.Encoder

type t

val create : int -> t

create capacity initializes an encoder with a dynamic table with maximum size capacity. This size is an approximation of the memory overhead in bytes.

See RFC75414.1 for more details.

val encode_header : t -> Faraday.t -> header -> unit

encode_header encoder f header writes an encoded header to the Faraday buffer f.

val set_capacity : t -> int -> unit

set_capacity encoder capacity sets encoder's dynamic table size to maximum size capacity. This size is an approximation of the memory overhead in bytes.

See RFC75406.5.2 and RFC75414.1 for more details.